Dont clear temp directory if running as teamcity

+++ Server/04-stop.ps1
@@ -14,9 +14,9 @@ TeamCity-Block("Stop and remove compose project") {
     write-host "Removing $ProjectName compose project"
     & docker-compose --file .\Server\docker-compose.yml --project-name $ProjectName down
 
-  & docker rm -f $ProjectName"_octopus_1"
+    & docker rm -f $ProjectName"_octopus_1"
 
-    if(Test-Path .\Temp) {
+    if(!Test-RunningUnderTeamCity -and (Test-Path .\Temp)) {
       Remove-Item .\Temp -Recurse -Force
     }
 }
